New table links and aliases component
