+++ src/scripts/modules/storage-explorer/react/modals/DeleteTableModal.jsx
@@ -1,7 +1,7 @@
 import React, { PropTypes } from 'react';
-import { Link } from 'react-router';
-import { Modal, Table, FormGroup, Checkbox } from 'react-bootstrap';
+import { Modal, FormGroup, Checkbox } from 'react-bootstrap';
 import ConfirmButtons from '../../../../react/common/ConfirmButtons';
+import TableAliasesAndLinks from '../components/TableAliasesAndLinks';
 
 export default React.createClass({
   propTypes: {
@@ -22,9 +22,6 @@ export default React.createClass({
   },
 
   render() {
-    const ownerId = this.props.sapiToken.getIn(['owner', 'id']);
-    const isOrganizationMember = this.props.sapiToken.getIn(['admin', 'isOrganizationMember']);
-
     return (
       <Modal show={this.props.show} onHide={this.props.onHide}>
         <Modal.Header closeButton>
@@ -43,60 +40,12 @@ export default React.createClass({
                 </Checkbox>
               </FormGroup>
 
-              <Table responsive striped>
-                <thead>
-                  <tr>
-                    <th>Table aliases</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {this.props.tableAliases.map(alias => (
-                    <tr key={alias.get('id')}>
-                      <td>
-                        <Link
-                          to="storage-explorer-table"
-                          params={{ bucketId: alias.getIn(['bucket', 'id']), tableName: alias.get('name') }}
-                          onClick={this.props.onHide}
-                        >
-                          {alias.get('id')}
-                        </Link>
-                      </td>
-                    </tr>
-                  ))}
-
-                  {this.props.tableLinks.map(alias => {
-                    const project = alias.get('project');
-
-                    return (
-                      <tr key={alias.get('id')}>
-                        <td>
-                          {ownerId === project.get('id') && (
-                            <Link
-                              to="storage-explorer-table"
-                              params={{ bucketId: alias.get('bucketId'), tableName: alias.get('tableName') }}
-                            >
-                              {alias.get('id')}
-                            </Link>
-                          )}
-                          {ownerId !== project.get('id') && isOrganizationMember && (
-                            <span>
-                              {project.get('name')} / {alias.get('id')}
-                            </span>
-                          )}
-                          {ownerId !== project.get('id') &&
-                            isOrganizationMember &&
-                            <a href={`/admin/projects/${project.get('id')}`}>{project.get('name')}</a> /
-                            (
-                              <a href={`/admin/projects/${project.get('id')}/storage#/buckets/${alias.get('id')}`}>
-                                {alias.get('id')}
-                              </a>
-                            )}
-                        </td>
-                      </tr>
-                    );
-                  })}
-                </tbody>
-              </Table>
+              <TableAliasesAndLinks
+                sapiToken={this.props.sapiToken}
+                tableAliases={this.props.tableAliases}
+                tableLinks={this.props.tableLinks}
+                onLinkClick={this.props.onHide}
+              />
             </div>
           )}
         </Modal.Body>
